Put episode rundown association code into its own concern and fragment cache some of the #published_content calls in the view.
