+++ app/models/content_shell.rb
@@ -18,6 +18,7 @@ class ContentShell < ActiveRecord::Base
   include Concern::Associations::MissedItContentAssociation
   include Concern::Associations::EditionsAssociation
   include Concern::Associations::VerticalArticleAssociation
+  include Concern::Associations::EpisodeRundownAssociation
   include Concern::Validations::PublishedAtValidation
   include Concern::Associations::PmpContentAssociation::StoryProfile
   #include Concern::Callbacks::CacheExpirationCallback
