+++ app/models/event.rb
@@ -16,6 +16,7 @@ class Event < ActiveRecord::Base
   include Concern::Associations::MissedItContentAssociation
   include Concern::Associations::VerticalArticleAssociation
   include Concern::Associations::ProgramArticleAssociation
+  include Concern::Associations::EpisodeRundownAssociation
   include Concern::Associations::PmpContentAssociation::StoryProfile
   include Concern::Callbacks::GenerateSlugCallback
   include Concern::Callbacks::GenerateTeaserCallback
