+++ app/models/pij_query.rb
@@ -15,6 +15,7 @@ class PijQuery < ActiveRecord::Base
   include Concern::Associations::MissedItContentAssociation
   include Concern::Associations::VerticalArticleAssociation
   include Concern::Associations::ProgramArticleAssociation
+  include Concern::Associations::EpisodeRundownAssociation
   include Concern::Validations::SlugValidation
   include Concern::Callbacks::GenerateSlugCallback
   include Concern::Model::Searchable
