+++ app/models/show_segment.rb
@@ -22,6 +22,8 @@ class ShowSegment < ActiveRecord::Base
   include Concern::Associations::EditionsAssociation
   include Concern::Associations::VerticalArticleAssociation
   include Concern::Associations::ProgramArticleAssociation
+  include Concern::Associations::PmpContentAssociation::StoryProfile
+  include Concern::Associations::EpisodeRundownAssociation
   include Concern::Validations::ContentValidation
   include Concern::Callbacks::SetPublishedAtCallback
   include Concern::Callbacks::GenerateSlugCallback
@@ -37,7 +39,6 @@ class ShowSegment < ActiveRecord::Base
   include Concern::Methods::CommentMethods
   include Concern::Methods::AssetDisplayMethods
   include Concern::Sanitizers::Content
-  include Concern::Associations::PmpContentAssociation::StoryProfile
 
   self.disqus_identifier_base = "shows/segment"
   self.public_route_key = "segment"
@@ -46,18 +47,6 @@ class ShowSegment < ActiveRecord::Base
   belongs_to :show,
     :class_name   => "KpccProgram",
     :touch        => true
-
-  has_many :rundowns,
-    :class_name     => "ShowRundown",
-    :foreign_key    => "content_id",
-    :as             => "content",
-    :dependent      => :destroy
-
-  has_many :episodes,
-    -> { order('status desc,air_date desc') },
-    :through    => :rundowns,
-    :source     => :episode,
-    :autosave   => true
 
   alias_attribute :url, :public_url
